move all aat files to att module and drop aat_ prefix
